refactor(model): extract denormalize planner to its own module
Phase 3 audit §5.2 — move the ~1100 LoC denormalization planner out
of DerivaModel into a focused module with a narrow consumer set.

WHY
====

DerivaModel is touched by every mixin in the codebase (1300 LoC of
catalog introspection — name_to_table, is_vocabulary, is_asset,
find_features, etc.). Until now it ALSO carried ~1100 LoC of
denormalization planner code (_build_join_tree, _prepare_wide_table,
sink-finding, path-ambiguity detection, FK-graph reachability
primitives). The two responsibilities have very different fan-out:

  * Introspection — consumed by every layer of the codebase
  * Planner — consumed by 4 sites total (local_db/denormalize.py,
    local_db/denormalizer.py, core/mixins/dataset.py, and a single
    line in dataset/dataset_bag.py)

Split per the fan-out: planner gets its own module, introspection
stays where every mixin already imports it.

WHAT MOVED
==========

src/deriva_ml/model/denormalize_planner.py (new, ~1100 LoC):

  * JoinNode dataclass + walk/walk_edges
  * denormalize_column_name(...) module helper
  * class DenormalizePlanner:
      - Reachability primitives:
          _is_topological_association  (was public; now private —
              had no real external consumer, just one internal
              denormalizer.py caller)
          _fk_neighbors
          _downstream_fk_sources
          _outbound_reachable
          _schema_to_paths
          _table_relationship
      - Planner rules (Rules 2/5/6 from the spec):
          _find_sinks
          _determine_row_per
          _enumerate_paths
          _find_path_ambiguities
      - Top-level entries:
          _build_join_tree
          _prepare_wide_table

DerivaModel keeps:

  * A _planner property (lazy DenormalizePlanner construction +
    caching on the instance).
  * A _schema_to_paths forwarder because dataset/dataset_bag.py
    still reaches it directly — flagged for a follow-up cleanup.
  * JoinNode + denormalize_column_name re-exported from catalog.py
    so existing imports keep working.

ALGORITHMIC OVERVIEW
====================

The module docstring now spells out the four questions the planner
answers (which table is the leaf?, is the leaf legal?, is the FK
path unique?, how do we JOIN?) and the three layers of methods that
compose into the answers. Class docstring points readers to the
canonical spec
(docs/superpowers/specs/2026-04-17-denormalization-semantics-design.md).

REVIEW-DRIVEN CLEANUPS
======================

In response to a senior-engineer review:

  * is_topological_association → _is_topological_association
    (was nominally public; had no external consumer — the audit's
    "Lens 4 rename" pre-staged but never landed).
  * Removed planner-local convenience accessors (ml_schema,
    domain_schemas, name_to_table, is_vocabulary, is_association,
    is_domain_schema). The planner now uses self.model.<attr>
    throughout — no double-API to maintain.
  * Stripped "Extracted from a nested function in _build_join_tree"
    historical notes from docstrings (the original extraction is
    no longer salient; the methods have a stable home now).
  * Replaced "legacy format" with "consumer format" in the
    _prepare_wide_table flatten step — there's no newer format;
    the "legacy" label was a fossil.
  * Replaced double-underscore _planner cache attribute name with
    single-underscore _planner_cache to avoid Python name-mangling
    making hasattr() spuriously False on first access.

EXTERNAL CALLSITE MIGRATION
===========================

~31 callsites in src/ and tests/ rewritten via
``model._foo(...)`` → ``model._planner._foo(...)`` (or
``self._model._foo(...)`` → ``self._model._planner._foo(...)``):

  src/deriva_ml/core/mixins/dataset.py     (1)
  src/deriva_ml/local_db/denormalize.py    (1)
  src/deriva_ml/local_db/denormalizer.py  (14)
  tests/dataset/test_composite_fk_denormalize.py  (1)
  tests/dataset/test_denormalize.py        (1)
  tests/local_db/test_denormalizer.py      (1)
  tests/local_db/test_planner_rules.py    (13)

NUMBERS
=======

  catalog.py:                2193 LoC → 836 LoC  (−1357)
  denormalize_planner.py:    new       → 1574 LoC

  Net: ~0 LoC moved; cleaner module boundary, narrower fan-out
  per file.

TEST RESULTS
============

  * tests/local_db/                       236 passed, 3 skipped
  * tests/asset/ + tests/model/ +
    tests/dataset/test_dataset_version_unit.py:
                                          458 passed, 3 skipped
  * Lint (ruff): clean.
