+++ src/deriva_ml/local_db/paged_fetcher_ermrest.py
@@ -25,6 +25,7 @@ from __future__ import annotations
 
 from typing import Any
 from urllib.parse import quote
+
 from deriva_ml.core.logging_config import get_logger
 
 logger = get_logger(__name__)
