+++ src/deriva_ml/local_db/schema.py
@@ -16,13 +16,13 @@ import sqlite3
 from pathlib import Path
 from typing import Any
 
+from deriva.bag import sqlite_helpers as sh
+from deriva.bag.schema import SchemaBuilder, SchemaORM
 from deriva.core.ermrest_model import Model
 from sqlalchemy import Table as SQLTable
 from sqlalchemy import event
 from sqlalchemy.engine import Engine
 
-from deriva.bag import sqlite_helpers as sh
-from deriva.bag.schema import SchemaBuilder, SchemaORM
 from deriva_ml.core.logging_config import get_logger
 
 logger = get_logger(__name__)
